Implement Pause menu

+++ source/HitObjects/HitCircle.cpp
@@ -5,7 +5,7 @@ HitCircle::HitCircle(int32_t x, int32_t y, int32_t time, HitObjectType type, Hit
 	uint32_t size = DifficultyManager::GetCircleSize();
 	
 	uint32_t preempt = DifficultyManager::GetPreemptTime();
-	int32_t fadeInStart = time - preempt;
+	int32_t fadeInStart = (int32_t)time - (int32_t)preempt;
 	int32_t fadeInEnd = fadeInStart + (preempt >> 3);
 	mEndTime = time + DifficultyManager::GetHitWindow50();
